[ActionSheet] Remove Beta test from Ready component (#6677)
## Related links
* Component: [ActionSheet](https://github.com/material-components/material-components-ios/tree/develop/components/ActionSheet)

## Introduction
When I moved `MDCActionSheet` to ready I didn't properly update the cocoapods tests. The test were relying on the `MaterialComponentsBeta` podspec. This change fixes their reliance on the `Beta` pod and the test weren't being run.

## The fix
This fixes the issue where the test weren't part of the scheme and relying on beta things.
