+++ components/ActionSheet/BUILD
@@ -117,6 +117,19 @@ swift_library(
     ],
     deps = [
         ":ActionSheet",
+        ":private",
+    ],
+)
+
+swift_library(
+    name = "unit_test_color_themer_swift_sources",
+    srcs = glob(["tests/unit/ColorThemer/*.swift"]),
+    copts = [
+        "-swift-version",
+        "3",
+    ],
+    deps = [
+        ":ActionSheet",
         ":ColorThemer",
         ":private",
     ],
@@ -125,9 +138,28 @@ swift_library(
 mdc_objc_library(
     name = "unit_test_sources",
     testonly = 1,
+    srcs = native.glob(["tests/unit/*.m",]),
+    hdrs = native.glob(["tests/unit/*.h",]),
+    sdk_frameworks = [
+        "UIKit",
+        "CoreImage",
+        "XCTest",
+    ],
+    deps = [
+        ":ActionSheet",
+        ":ActionSheetThemer",
+        ":ColorThemer",
+        ":Theming",
+        ":TypographyThemer",
+        ":privateHeaders",
+    ],
+)
+
+mdc_objc_library(
+    name = "unit_test_theming_sources",
+    testonly = 1,
     srcs = native.glob([
-        "tests/unit/*.m",
-        "tests/unit/*.h",
+        "tests/unit/Theming/*.m",
     ]),
     sdk_frameworks = [
         "UIKit",
@@ -141,6 +173,28 @@ mdc_objc_library(
         ":Theming",
         ":TypographyThemer",
         ":privateHeaders",
+        ":unit_test_sources",
+    ],
+)
+
+mdc_objc_library(
+    name = "unit_test_themer_sources",
+    testonly = 1,
+    srcs = native.glob([
+        "tests/unit/ActionSheetThemer/*.m",
+    ]),
+    sdk_frameworks = [
+        "UIKit",
+        "CoreImage",
+        "XCTest",
+    ],
+    deps = [
+        ":ActionSheet",
+        ":ActionSheetThemer",
+        ":ColorThemer",
+        ":TypographyThemer",
+        ":privateHeaders",
+        ":unit_test_sources",
     ],
 )
