+++ components/ActionSheet/tests/unit/ActionSheetTest.swift
@@ -14,7 +14,6 @@
 
 import XCTest
 import MaterialComponents.MaterialActionSheet
-import MaterialComponentsBeta.MaterialActionSheet_ColorThemer
 
 class ActionSheetTest: XCTestCase {
 
@@ -108,20 +107,5 @@ class ActionSheetTest: XCTestCase {
     // Then
     XCTAssertEqual(actionSheet.view.backgroundColor, actionSheet.backgroundColor)
     XCTAssertEqual(actionSheet.view.backgroundColor, newBackgroundColor)
-  }
-
-  func testApplyThemerToBackgroundColor() {
-    // Given
-    let surfaceColor: UIColor = .blue
-    let colorScheme = MDCSemanticColorScheme()
-    colorScheme.surfaceColor = surfaceColor
-
-    // When
-    MDCActionSheetColorThemer.applySemanticColorScheme(colorScheme, to: actionSheet)
-    let _ = actionSheet.view
-
-    // Then
-    XCTAssertEqual(actionSheet.view.backgroundColor, surfaceColor)
-    XCTAssertEqual(actionSheet.backgroundColor, surfaceColor)
   }
 }
