+++ components/ActionSheet/tests/unit/ColorThemer/MDCActionSheetColorThemerTest.swift
@@ -1,0 +1,36 @@
+// Copyright 2019-present the Material Components for iOS authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import XCTest
+import MaterialComponents.MaterialActionSheet
+import MaterialComponentsBeta.MaterialActionSheet_ColorThemer
+
+class ActionSheetColorThemerTest: XCTestCase {
+  func testApplyThemerToBackgroundColor() {
+    // Given
+    let actionSheet = MDCActionSheetController()
+    let surfaceColor: UIColor = .blue
+    let colorScheme = MDCSemanticColorScheme()
+    colorScheme.surfaceColor = surfaceColor
+
+    // When
+    MDCActionSheetColorThemer.applySemanticColorScheme(colorScheme, to: actionSheet)
+    let _ = actionSheet.view
+
+    // Then
+    XCTAssertEqual(actionSheet.view.backgroundColor, surfaceColor)
+    XCTAssertEqual(actionSheet.backgroundColor, surfaceColor)
+  }
+}
+
